This is a short POC to demonstrate how we could build a "recreatable" tenant cosmos container.
While the intended scenario is "key roatation" it woudl work for any case where e.g. configuration had changed and you wanted to recreate your connection.

I have implemented it over the ComsosContainer as an example.

In use, you would need to be aware of the possibility of recreation - this might mean, for example, that the Container/Client you are using is disposed from underneath you, and you would have to go and get it again; you cannot mix recreatable and non-recreatable users in the same environment.