Add tenanted name helper (#385)
This provides a public way to form tenanted container names in a standard way, and also the means to convert those into hashed names for use in Azure Blob Storage. The hashing part is now implemented on top of the container name hashing in Corvus.Storage.

Also took this opportunity to make ITenant.ETag read-only. the fact that it was writable was an oversight, and we're about to release v3.0.0, so this is a good time to fix that. (Tenant.ETag remains writable, and that's the one that needs to be.)