Add updated Azure Tables support (#409)
Also fixed bug when using logical container/table names. It turns out that neither the new Azure Tables support nor the old Azure Blob Storage support were correctly handling the case where the configuration did not contain the table name, and we relied on a logical name being passed in. The logical name itself was getting hashed when we should have been using a hash of the *tenanted* logical name.